pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-resource-optimizer"
version = "0.1.1"
description = "Cloud cost optimization CLI"
requires-python = ">=3.10"
dependencies = [
  "azure-identity>=1.16.0",
  "azure-storage-blob>=12.20.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
cloudopt = "cloudopt_cli.cli:main"
